+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.7.0 (unreleased)
 ------------------
 
+- #2805 Migrate Worksheets to DX
 - #2808 Sticky form tabbing
 - #2807 Fix traceback when emailing results report and email 'From' not set
 - #2806 Allow to create sample partitions on copy